Upgrade both input and loop counter to long long

+++ sum_calculator.cpp
@@ -2,13 +2,13 @@
 using namespace std;
 
 int main() {
-    int number = 0;
+    long long number = 0;
     long long sum = 0;
     
     cout << "Enter number: ";
     cin >> number;
     
-    for (int i = 1; i <= number; i++) {
+    for (long long i = 1; i <= number; i++) {
         
         sum = sum + i;
     }
